aiijcBabyshark/flask app/requirements.txt
absl-py==0.13.0
astunparse==1.6.3
autopep8==1.5.4
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
clang==5.0
click==8.0.1
colorama==0.4.4
Flask==2.0.1
flatbuffers==1.12
gast==0.4.0
# google-auth==1.35.0
# google-auth-oauthlib==0.4.6
# google-pasta==0.2.0
grpcio==1.39.0
gunicorn==20.1.0
h5py==3.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
keras==2.6.0
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==2.0.1
numpy==1.19.5
oauthlib==3.1.1
opencv-python-headless==4.5.3.56
opt-einsum==3.3.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
python-dateutil==2.8.2
pytz==2021.1
# requests==2.26.0
# requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
toml==0.10.2
typing-extensions==3.7.4.3
Werkzeug==2.0.1
wincertstore==0.2
wrapt==1.12.1
